WIP: HWS: constrain rule link list chasing
FIXME: This negatively interacts wtih table resizing, so needs
additional work.

The current STE-linked-list chasing logic terminates whenever it reaches
an STE address that is not a rule STE included in the current dump. This
works fine when the chain of STEs reaches an anchor STE or an STE truly
outside the "program" being dumped. However, if STE content is optimized
so that some STEs jump directly to another table's rule STEs rather than
to that table's anchor STE, the current algorithm will dump too many
STEs. Enhance the code so that it only considers rule STEs within the
current matcher to be non-terminating. Thus, each jump between tables
will be explicitly visible.